Updates to VS Code intellisense, build setup and speedup
This PR does a few combined things. First it enables C++ intellisense in vscode in addition to the java intellisense.

Second it uses the proper extracted cross compile JNI setup, which is more reliable and easier. Youll
notice intellisense will be broken at first because of this in C++. Run a build, it will fix itself.

Finally, it switches the JNI code to not pass the byte array (This resulted in 3 extra full copies of the image across memory),
and instead does this all in C++. It passes a created CameraServer handle down through JNI, so that still happens in java.